requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
certifi==2019.6.16
Django==2.2.4
django-crispy-forms==1.7.2
flake8==3.7.8
mysqlclient==1.4.2.post1
pytz==2019.2
sqlparse==0.3.0
docker==2.6.1
docker-compose==1.17.1
requests==2.11.1
whitenoise==4.1.3